Add `force-import' macro
When writing modules, you usually limit the visibility of symbols to
those that the users of the module need to use it. This is done to avoid
littering the global name-space.

Now with test-suites, you will very likely want to test internal
functionality, as well. To do that, you might use guile's `@@' function
to access non-exported symbols from a given module:

  (define node? (@@ (data tree) node?)

The `force-import' macro makes the following:

  (force-import (data tree) node? node-children node-parent)

equivalent to:

  (define node? (@@ (data tree) node?)
  (define node-children (@@ (data tree) node-children)
  (define node-parent (@@ (data tree) node-parent)

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>