Rework expression handling entirely
The system should be quite graceful towards what an expression may
cause. For example, after this:

  (define-tap-test (pass-if-= a b) (= a b))

...consider this:

  (pass-if-= (something 123) 23)

In this case (something 123) may well throw an exception. That would be
unacceptable with respect to the result of the computation, but it
should never be able to break the run of the rest of a test-suite.

The obvious solution is to put a catch around the final expression that
is assembled, which works well enough, but makes it hard to generate
good diagnostics about what actually happened.

This new approach works like this:

  - First evaluate each input expression (a and b in the above
    example) and store it in a temporary variable, while also
    nothing whether or not the expression threw an exception
    along the way.

  - Take the expression ("(= a b)" in the example) and replace
    each and every occurance of "a" and "b" with the value
    retrieved from the evaluation of these expression from
    step 1.

  - This expression can now be evaluated again and can also be
    safely wrapped in yet another catch expression.

This allows for graceful behaviour towards expected and unexpected
exceptions during expression evaluation as well as full access to all
the information about any failures or exceptions, which allows for high
level diagnostics.

The new code passes the current-test suite, but output, formatting and
diagnostics still need to be improved.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>