Export ‘define-tap-test’ macro
This allows the user to define custom test macros, to test common
situations with a given code-base, that is not covered by the core
framework.

This used to be possible, wrapping the very generic pass-if-true and
pass-if-false macros within a custom macro, but this allows the test
definition to be much more concise and also gives the full range of all
diagnostics that the framework supports to custom tests.

Example:

(define-tap-test (pass-if-valid-type? datum) (or (string? datum)
                                                 (number? datum)))

(define-test "Costum macro pass-if-valid-type? likes strings"
  (pass-if-valid-type? "123"))

=> ok 1 - Costum macro pass-if-valid-type? likes strings

(define-test "Costum macro pass-if-valid-type? likes numbers"
  (pass-if-valid-type? 4+2i))

=> ok 2 - Costum macro pass-if-valid-type? likes numbers

(define-test "Costum macro pass-if-valid-type? likes lists"
  (pass-if-valid-type? '(a b c)))

=> not ok 3 - Costum macro pass-if-valid-type? likes lists
   #
   # failed test: "Costum macro pass-if-valid-type? likes lists"
   #
   # location:
   #     file: "some-test-file.t"
   #     line: 125
   #   column: 4
   #
   # full test-expression:
   #     (pass-if-valid-type? '(a b c))
   #
   # form:
   #     (pass-if-valid-type? datum)
   #
   # expression:
   #     (or (string? datum) (number? datum))
   #
   # evaluated form:
   #     (pass-if-valid-type? (a b c))
   #
   # evaluated expression:
   #     (or (string? (a b c)) (number? (a b c)))
   #

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>